pyproject.toml
[tool.poetry]
name = "googlyify"
version = "0.1.0"
description = "Draw googly eyes on people with Amazon Rekognition"
authors = ["James Seward <sewardjm@amazon.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"
typer = "^0.4.0"
boto3 = "^1.21.0"

[tool.poetry.dev-dependencies]
types-boto3 = {extras = ["rekognition"], version = "^1.0.1"}
pylint = "^2.12.2"
mypy = "^0.931"
types-Pillow = "^9.0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
